* HDDS-4944: (268 commits)
  HDDS-6366. [Multi-Tenant] Disallow specifying custom accessId in OzoneManager (#3166)
  HDDS-6275. [Multi-Tenant] Add feature documentation and CLI quick start guide (#3095)
  HDDS-6063. [Multi-Tenant] Use VOLUME_LOCK in read and write requests, and some minor refactoring (#3051)
  HDDS-6214. [Multi-Tenant] Fix KMS Encryption/Decryption (#3010)
  HDDS-6322. Fix Recon getting inccorrect sequenceNumber from OM (#3090)
  HDDS-5913. Avoid integer overflow when setting dfs.container.ratis.lo… (#2785)
  HDDS-6313. Remove replicas in ContainerStateMap when a container is deleted (#3086)
  HDDS-6186. Selective checks: skip integration check for unit test changes (#3061)
  HDDS-6310. Update json-smart to 2.4.7. (#3080)
  HDDS-6190. Cleanup unnecessary datanode id path checks. (#2993)
  HDDS-6304. Add enforcer to make sure ozone.version equals project.version (#3075)
  HDDS-6309. Update ozone-runner version to 20220212-1 (#3079)
  HDDS-6293. Allow using custom ozone-runner image (#3072)
  HDDS-4126. Freon key generator should support >2GB files. (#3054)
  HDDS-6088. Implement O3FS/OFS getFileChecksum() using file checksum helpers - addendum: fix checkstyle
  HDDS-6088. Implement O3FS/OFS getFileChecksum() using file checksum helpers. (#2935)
  HDDS-6084. [Multi-Tenant] Handle upgrades to version supporting S3 multi-tenancy (#3018)
  HDDS-6257. Wrong stack trace for S3 errors (#3066)
  HDDS-6278 Improve memory profile for listStatus API call. (#3053)
  HDDS-6285. ozonesecure-mr intermittently failing with timeout downloading packages (#3057)
  ...

+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/freon/HadoopDirTreeGenerator.java
@@ -70,7 +70,7 @@ public class HadoopDirTreeGenerator extends BaseFreonGenerator
           "written in each directory. Full name --fileSize will be removed " +
           "in later versions.",
       defaultValue = "4096")
-  private int fileSizeInBytes;
+  private long fileSizeInBytes;
 
   @Option(names = {"-b", "--buffer"},
           description = "Size of buffer used to generated the file content.",
@@ -175,7 +175,7 @@ public class HadoopDirTreeGenerator extends BaseFreonGenerator
       }
     }
 
-    while(spanIndex < span) {
+    while (spanIndex < span) {
       String levelSubDir = makeDirWithGivenNumberOfFiles(parent);
       ++spanIndex;
